[TASK] split code
