+++ custom_components/better_thermostat/climate.py
@@ -1,63 +1,34 @@
 """Better Thermostat"""
 
 import asyncio
-import json
 import logging
-import math
 import numbers
-from abc import ABC
-from datetime import datetime, timedelta
-from random import randint
-from time import sleep
-
 import homeassistant.helpers.config_validation as cv
 import homeassistant.util.dt as dt_util
 import voluptuous as vol
+from abc import ABC
+
+from datetime import datetime, timedelta
+from random import randint
 from homeassistant.components.climate import ClimateEntity, PLATFORM_SCHEMA
-from homeassistant.components.climate.const import (
-	CURRENT_HVAC_HEAT, CURRENT_HVAC_IDLE, CURRENT_HVAC_OFF, HVAC_MODE_HEAT, HVAC_MODE_OFF, SUPPORT_TARGET_TEMPERATURE
-)
-from homeassistant.components.recorder import history
-from homeassistant.const import (ATTR_TEMPERATURE, CONF_NAME, CONF_UNIQUE_ID, EVENT_HOMEASSISTANT_START, STATE_UNAVAILABLE, STATE_UNKNOWN)
+from homeassistant.components.climate.const import (CURRENT_HVAC_HEAT, CURRENT_HVAC_IDLE, CURRENT_HVAC_OFF, HVAC_MODE_HEAT, HVAC_MODE_OFF)
+from homeassistant.const import (ATTR_TEMPERATURE, CONF_NAME, CONF_UNIQUE_ID, EVENT_HOMEASSISTANT_START)
 from homeassistant.core import callback, CoreState
-from homeassistant.helpers.entity_registry import (async_entries_for_config_entry)
 from homeassistant.helpers.event import (async_track_state_change_event, async_track_time_change)
 from homeassistant.helpers.reload import async_setup_reload_service
 from homeassistant.helpers.restore_state import RestoreEntity
-
 from . import DOMAIN, PLATFORMS
-from .helpers import check_float, set_trv_values
-from .models.models import convert_inbound_states, convert_outbound_states, get_device_model
+from .const import (VERSION,SUPPORT_FLAGS,CONF_HEATER,CONF_SENSOR,CONF_SENSOR_WINDOW,CONF_WEATHER,CONF_OUTDOOR_SENSOR,CONF_OFF_TEMPERATURE,CONF_WINDOW_TIMEOUT,CONF_VALVE_MAINTENANCE,CONF_NIGHT_TEMP,CONF_NIGHT_START,CONF_NIGHT_END,CONF_MIN_TEMP,CONF_MAX_TEMP,CONF_PRECISION,CONF_TARGET_TEMP,ATTR_STATE_CALL_FOR_HEAT,ATTR_STATE_DAY_SET_TEMP,ATTR_STATE_LAST_CHANGE,ATTR_STATE_NIGHT_MODE,ATTR_STATE_WINDOW_OPEN,ATTR_VALVE_POSITION,DEFAULT_NAME)
+from .helpers import check_float, startup
+from .models.models import convert_outbound_states, get_device_model
+from .controlling import set_hvac_mode, set_target_temperature, set_trv_values, trv_valve_maintenance
+from .events.temperature import trigger_temperature_change
+from .events.time import trigger_time
+from .events.trv import trigger_trv_change
+from .events.window import check_window_state, trigger_window_change
+from .weather import check_summer_state, check_weather
 
 _LOGGER = logging.getLogger(__name__)
-
-DEFAULT_NAME = "Better Thermostat"
-
-CONF_HEATER = "thermostat"
-CONF_SENSOR = "temperature_sensor"
-CONF_SENSOR_WINDOW = "window_sensors"
-CONF_TARGET_TEMP = "target_temp"
-CONF_WEATHER = "weather"
-CONF_OFF_TEMPERATURE = "off_temperature"
-CONF_WINDOW_TIMEOUT = "window_off_delay"
-CONF_OUTDOOR_SENSOR = "outdoor_sensor"
-CONF_VALVE_MAINTENANCE = "valve_maintenance"
-CONF_NIGHT_TEMP = "night_temp"
-CONF_NIGHT_START = "night_start"
-CONF_NIGHT_END = "night_end"
-CONF_MIN_TEMP = "min_temp"
-CONF_MAX_TEMP = "max_temp"
-CONF_PRECISION = "precision"
-
-SUPPORT_FLAGS = SUPPORT_TARGET_TEMPERATURE
-
-ATTR_STATE_WINDOW_OPEN = "window_open"
-ATTR_STATE_NIGHT_MODE = "night_mode"
-ATTR_STATE_CALL_FOR_HEAT = "call_for_heat"
-ATTR_STATE_LAST_CHANGE = "last_change"
-ATTR_STATE_DAY_SET_TEMP = "last_day_set_temp"
-
-ATTR_VALVE_POSITION = "valve_position"
 
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(
 	{
@@ -205,13 +176,7 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 		self._state_class = state_class
 		self.local_temperature_calibration_entity = None
 		self.valve_position_entity = None
-		try:
-			with open("/config/custom_components/better_thermostat/manifest.json") as manifest_file:
-				manifest = json.load(manifest_file)
-				self.version = manifest["version"]
-		except (FileNotFoundError, KeyError, json.JSONDecodeError) as e:
-			_LOGGER.error("better_thermostat %s: could not read version from manifest file - %s", not self._name, e)
-			self.version = "unknown"
+		self.version = VERSION
 		self.last_change = None
 		self.load_saved_state = False
 		self._last_reported_valve_position = None
@@ -229,14 +194,14 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 
 		# Add listener
 		async_track_state_change_event(
-			self.hass, [self.sensor_entity_id], self._async_sensor_changed
+			self.hass, [self.sensor_entity_id], trigger_temperature_change
 		)
 		async_track_state_change_event(
-			self.hass, [self.heater_entity_id], self._async_trv_changed
+			self.hass, [self.heater_entity_id], trigger_trv_change
 		)
 		if self.window_sensors_entity_ids:
 			async_track_state_change_event(
-				self.hass, [self.window_sensors_entity_ids], self._async_window_changed
+				self.hass, [self.window_sensors_entity_ids], trigger_window_change
 			)
 		
 		# check if night mode was configured
@@ -244,14 +209,14 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 			_LOGGER.debug("Night mode configured")
 			async_track_time_change(
 				self.hass,
-				self._async_timer_trigger,
+				trigger_time,
 				self.night_start.hour,
 				self.night_start.minute,
 				self.night_start.second,
 			)
 			async_track_time_change(
 				self.hass,
-				self._async_timer_trigger,
+				trigger_time,
 				self.night_end.hour,
 				self.night_end.minute,
 				self.night_end.second,
@@ -264,7 +229,7 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 			_LOGGER.info("better_thermostat %s: Starting version %s. Waiting for entity to be ready...", self.name, self.version)
 			
 			loop = asyncio.get_event_loop()
-			loop.create_task(self.startup())
+			loop.create_task(startup())
 		
 		if self.hass.state == CoreState.running:
 			_async_startup()
@@ -357,115 +322,7 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 			self._hvac_mode = HVAC_MODE_OFF
 		self.async_write_ha_state()
 
-	async def startup(self):
-		"""Run startup tasks."""
-		window = None
-		await asyncio.sleep(5)
-		
-		while self.startup_running:
-			sensor_state = self.hass.states.get(self.sensor_entity_id)
-			trv_state = self.hass.states.get(self.heater_entity_id)
-			
-			if sensor_state is None:
-				_LOGGER.error(
-					"better_thermostat %s: configured temperature sensor with id '%s' could not be located",
-					self.name,
-					self.sensor_entity_id
-				)
-				return False
-			if trv_state is None:
-				_LOGGER.error(
-					"better_thermostat %s: configured TRV/climate entry with id '%s' could not be located",
-					self.name,
-					self.heater_entity_id
-				)
-				return False
-			if self.window_sensors_entity_ids:
-				window = self.hass.states.get(self.window_sensors_entity_ids)
-				
-				if window is None:
-					_LOGGER.error(
-						"better_thermostat %s: configured window sensor entry or group with id '%s' could not be located",
-						self.name,
-						self.window_sensors_entity_ids
-					)
-					return False
-				
-				# make sure window has a state variable
-				try:
-					if window.state:
-						pass
-				except (ValueError, NameError, AttributeError):
-					_LOGGER.error(
-						"better_thermostat %s: configured window sensor entry or group with id '%s' could not be located",
-						self.name,
-						self.window_sensors_entity_ids
-					)
-					return False
-			
-			_ready = True
-			
-			if sensor_state.state in (STATE_UNAVAILABLE, STATE_UNKNOWN, None):
-				_LOGGER.info(
-					"better_thermostat %s: waiting for sensor entity with id '%s' to become fully available...",
-					self.name,
-					self.sensor_entity_id
-				)
-				_ready = False
-			if trv_state.state in (STATE_UNAVAILABLE, STATE_UNKNOWN, None):
-				_LOGGER.info(
-					"better_thermostat %s: waiting for TRV/climate entity with id '%s' to become fully available...",
-					self.name,
-					self.heater_entity_id
-				)
-				_ready = False
-			
-			if self.window_sensors_entity_ids in (STATE_UNAVAILABLE, STATE_UNKNOWN, None) or window.state in (
-					STATE_UNAVAILABLE, STATE_UNKNOWN, None):
-				_LOGGER.info(
-					"better_thermostat %s: waiting for window sensor entity with id '%s' to become fully available...",
-					self.name,
-					self.window_sensors_entity_ids
-				)
-				_ready = False
-			
-			if not _ready:
-				_LOGGER.info("better_thermostat %s: retry startup in 15 seconds...", self.name)
-				await asyncio.sleep(15)
-				continue
-			
-			if self.window_sensors_entity_ids:
-				window = self.hass.states.get(self.window_sensors_entity_ids)
-				
-				check = window.state
-				if check == 'on':
-					self.window_open = True
-					self._hvac_mode = HVAC_MODE_OFF
-				else:
-					self.window_open = False
-					self.closed_window_triggered = False
-				_LOGGER.debug(
-					"better_thermostat %s: detected window state at startup: %s",
-					self.name,
-					"Open" if self.window_open else "Closed"
-				)
-			
-			self.startup_running = False
-			entity_registry = await self.hass.helpers.entity_registry.async_get_registry()
-			reg_entity = entity_registry.async_get(self.heater_entity_id)
-			entity_entries = async_entries_for_config_entry(entity_registry, reg_entity.config_entry_id)
-			for entity in entity_entries:
-				uid = entity.unique_id
-				if "local_temperature_calibration" in uid:
-					self.local_temperature_calibration_entity = entity.entity_id
-				if "valve_position" in uid:
-					self.valve_position_entity = entity.entity_id
-			self._async_update_temp(sensor_state)
-			self.async_write_ha_state()
-			await asyncio.sleep(5)
-			_LOGGER.info("better_thermostat %s: startup completed.", self.name)
-			await self._async_control_heating()
-		return True
+
 	
 	@property
 	def extra_state_attributes(self):
@@ -479,43 +336,6 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 		}
 		
 		return dev_specific
-	
-	@callback
-	def _nighttime(self, current_time):
-		"""
-		Return whether it is nighttime.
-		@param current_time: time.time()
-		@return: bool True if it is nighttime; None if not configured
-		"""
-		_return_value = None
-		
-		# one or more of the inputs is None or empty
-		if not all([self.night_start, self.night_end, current_time]):
-			return _return_value
-		
-		# fetch to instance variables, since we might want to swap them
-		start_time, end_time = self.night_start, self.night_end
-		
-		# if later set to true we'll swap the variables and output boolean, 'cause we use the logic backwards
-		#   if the nighttime passes not over midnight, like (01:00 to 05:00) we use the inverted logic
-		#   while something like 23:00 to 05:00 would use the default
-		_reverse = False
-		
-		if start_time.hour < end_time.hour or (start_time.hour == end_time.hour and start_time.minute < end_time.minute):
-			# not passing midnight, so we use the inverted logic
-			_reverse = True
-			start_time, end_time = end_time, start_time
-		
-		# if we are after the start time, but before the end time, we are in the night
-		if (current_time.hour > start_time.hour or (
-				current_time.hour == start_time.hour and current_time.minute >= start_time.minute)) and current_time.hour < end_time.hour or (
-				current_time.hour == end_time.hour and current_time.minute < end_time.minute):
-			_return_value = True
-		
-		# flip output, since we flipped the start/end time
-		if _reverse:
-			return not _return_value
-		return _return_value
 	
 	@property
 	def available(self):
@@ -617,23 +437,11 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 	
 	async def async_set_hvac_mode(self, hvac_mode):
 		"""Set hvac mode."""
-		if hvac_mode == HVAC_MODE_HEAT:
-			self._hvac_mode = HVAC_MODE_HEAT
-		elif hvac_mode == HVAC_MODE_OFF:
-			self._hvac_mode = HVAC_MODE_OFF
-		else:
-			_LOGGER.error("better_thermostat %s: Unsupported hvac_mode %s", self.name, hvac_mode)
-		self.async_write_ha_state()
-		await self._async_control_heating()
+		await set_hvac_mode(self, hvac_mode)
 	
 	async def async_set_temperature(self, **kwargs):
 		"""Set new target temperature."""
-		temperature = kwargs.get(ATTR_TEMPERATURE)
-		if temperature is None:
-			return
-		self._target_temp = temperature
-		self.async_write_ha_state()
-		await self._async_control_heating()
+		await set_target_temperature(self, **kwargs)
 	
 	@property
 	def min_temp(self):
@@ -652,251 +460,6 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 		
 		# Get default temp from super class
 		return super().max_temp
-	
-	@callback
-	async def _async_timer_trigger(self, current_time):
-		"""
-		Triggered by night mode timer.
-		@param current_time:
-		"""
-		
-		_is_night = self._nighttime(current_time)
-		
-		if _is_night is None:
-			_LOGGER.error("better_thermostat %s: Error while checking if it is night", self.name)
-			return
-		elif _is_night:
-			_LOGGER.debug("better_thermostat %s: Night mode activated", self.name)
-			self.last_daytime_temp = self._target_temp
-			self._target_temp = self.night_temp
-			self.night_mode_active = True
-		
-		else:
-			_LOGGER.debug("ai_thermostat %s: Day mode activated", self.name)
-			if self.last_daytime_temp is None:
-				_LOGGER.error("better_thermostat %s: Could not load last daytime temp; continue using the current setpoint", self.name)
-			else:
-				self._target_temp = self.last_daytime_temp
-			self.night_mode_active = False
-		
-		self.async_write_ha_state()
-		await self._async_control_heating()
-	
-	@callback
-	async def _async_window_changed(self, event):
-		if self.startup_running:
-			return
-		if self.hass.states.get(self.heater_entity_id) is not None:
-			await asyncio.sleep(int(self.window_delay))
-			check = self.hass.states.get(self.window_sensors_entity_ids).state
-			if check == 'on':
-				self.window_open = True
-			else:
-				self.window_open = False
-			_LOGGER.debug("better_thermostat %s: Window (group) state changed to %s", self.name, "open" if self.window_open else "closed")
-			self.async_write_ha_state()
-			await self._async_control_heating()
-	
-	@callback
-	async def _async_sensor_changed(self, event):
-		"""Handle temperature changes."""
-		if self.startup_running:
-			return
-		new_state = event.data.get("new_state")
-		if new_state is None or new_state.state in (STATE_UNAVAILABLE, STATE_UNKNOWN):
-			return
-		
-		self._async_update_temp(new_state)
-		self.async_write_ha_state()
-		await self._async_control_heating()
-	
-	@callback
-	def _async_update_temp(self, state):
-		"""Update thermostat with the latest state from sensor."""
-		try:
-			self._cur_temp = float(state.state)
-		except (ValueError, AttributeError, KeyError, TypeError, NameError, IndexError):
-			_LOGGER.error(
-				"better_thermostat %s: Unable to update temperature sensor status from status update, current temperature not a number",
-				self.name
-			)
-	
-	@callback
-	async def _async_trv_changed(self, event):
-		"""Process TRV status updates"""
-		if self.startup_running:
-			return
-		
-		old_state = event.data.get("old_state")
-		new_state = event.data.get("new_state")
-		
-		if new_state is None or old_state is None:
-			return
-		
-		if new_state.attributes is not None:
-			try:
-				remapped_state = convert_inbound_states(self, new_state.attributes)
-				
-				# write valve position to local variable
-				if remapped_state.get(ATTR_VALVE_POSITION) is not None:
-					self._last_reported_valve_position = remapped_state.get(ATTR_VALVE_POSITION)
-					self._last_reported_valve_position_update_wait_lock.release()
-				
-				if old_state.attributes.get('system_mode') != new_state.attributes.get('system_mode'):
-					self._hvac_mode = remapped_state.get('system_mode')
-					
-					if self._hvac_mode != HVAC_MODE_OFF and self.window_open:
-						self._hvac_mode = HVAC_MODE_OFF
-						_LOGGER.debug("better_thermostat %s: Window open, turn off the heater", self.name)
-						await self._async_control_heating()
-				
-				if not self.ignore_states and new_state.attributes.get(
-						'current_heating_setpoint'
-				) is not None and self._hvac_mode != HVAC_MODE_OFF and self.calibration_type == 0:
-					_new_heating_setpoint = float(new_state.attributes.get('current_heating_setpoint'))
-					# if new setpoint is lower than the min setpoint, set it to the min setpoint
-					_overwrite_thermostat_update = False
-					if _new_heating_setpoint < self._min_temp:
-						_new_heating_setpoint = self._min_temp
-						_overwrite_thermostat_update = True
-					# if new setpoint is higher than the max setpoint, set it to the max setpoint
-					if _new_heating_setpoint > self._max_temp:
-						_new_heating_setpoint = self._max_temp
-						_overwrite_thermostat_update = True
-					if self._target_temp != _new_heating_setpoint:
-						self._target_temp = _new_heating_setpoint
-					else:
-						_overwrite_thermostat_update = False
-					
-					# if the user has changed the setpoint to a value that is not in the allowed range,
-					# overwrite the change with an TRV update
-					if _overwrite_thermostat_update:
-						_LOGGER.warning(
-							"better_thermostat %s: Overwriting setpoint %s with %s, as the new setpoint is out of bound (min/max temperature)",
-							self.name,
-							new_state.attributes.get('current_heating_setpoint'),
-							_new_heating_setpoint
-						)
-						await self._async_control_heating()
-			
-			
-			except TypeError as e:
-				_LOGGER.debug("better_thermostat entity not ready or device is currently not supported %s", e)
-			
-			self.async_write_ha_state()
-	
-	async def trv_valve_maintenance(self):
-		"""Maintenance of the TRV valve."""
-		
-		_LOGGER.info("better_thermostat %s: maintenance started", self.name)
-		
-		self.ignore_states = True
-		
-		if self.model == "TS0601_thermostat":
-			_LOGGER.debug("better_thermostat %s: maintenance will run TS0601_thermostat variant of cycle", self.name)
-			
-			# get current HVAC mode from HA
-			try:
-				_last_hvac_mode = self.hass.states.get(self.heater_entity_id).state
-			except:
-				_LOGGER.error("better_thermostat %s: Could not load current HVAC mode", self.name)
-				self.ignore_states = False
-				return
-			
-			_i = 0
-			_retry_limit_reached = False
-			
-			while True:
-				# close valve
-				_set_HVAC_mode_retry = 0
-				while not self._last_reported_valve_position == 0:
-					# send close valve command and wait for the valve to close
-					await set_trv_values(self, 'system_mode', 'off')
-					# wait for an update by the TRV on the valve position
-					await self._last_reported_valve_position_update_wait_lock.acquire()
-					if not self._last_reported_valve_position == 0 and _set_HVAC_mode_retry < 3:
-						_set_HVAC_mode_retry += 1
-						continue
-					elif _set_HVAC_mode_retry == 3:
-						_LOGGER.error("better_thermostat %s: maintenance could not close valve after 3 retries", self.name)
-						_retry_limit_reached = True
-						break
-					# wait 60 seconds to not overheat the motor
-					await asyncio.sleep(60)
-				
-				if _retry_limit_reached:
-					_LOGGER.error("better_thermostat %s: maintenance was aborted prematurely due to errors", self.name)
-					break
-				
-				# end loop after 3 opening cycles
-				elif _i > 3:
-					_LOGGER.info("better_thermostat %s: maintenance completed", self.name)
-					break
-				
-				# open valve
-				_set_HVAC_mode_retry = 0
-				while not self._last_reported_valve_position == 100:
-					# send open valve command and wait for the valve to open
-					await self.hass.services.async_call(
-						'climate', SERVICE_SET_HVAC_MODE, {'entity_id': self.heater_entity_id, 'hvac_mode': 'heat'}, blocking=True
-					)
-					await self._last_reported_valve_position_update_wait_lock.acquire()
-					if not self._last_reported_valve_position == 0 and _set_HVAC_mode_retry < 3:
-						_set_HVAC_mode_retry += 1
-						continue
-					elif _set_HVAC_mode_retry == 3:
-						_LOGGER.error("better_thermostat %s: maintenance could not open valve after 3 retries", self.name)
-						_retry_limit_reached = True
-						break
-					# wait 60 seconds to not overheat the motor
-					await asyncio.sleep(60)
-				
-				if _retry_limit_reached:
-					_LOGGER.error("better_thermostat %s: maintenance was aborted prematurely due to errors", self.name)
-					break
-				
-				_i += 1
-			
-			# returning the TRV to the previous HVAC mode
-			await self.hass.services.async_call(
-				'climate', SERVICE_SET_HVAC_MODE, {'entity_id': self.heater_entity_id, 'hvac_mode': _last_hvac_mode}, blocking=True
-			)
-			# give the TRV time to process the mode change and report back to HA
-			await asyncio.sleep(120)
-		
-		else:
-			
-			valve_position_available = False
-			# check if there's a valve_position field
-			try:
-				self.hass.states.get(self.heater_entity_id).attributes.get('valve_position')
-				valve_position_available = True
-			except:
-				pass
-			
-			if valve_position_available:
-				await set_trv_values(self, 'valve_position', 255)
-				await asyncio.sleep(60)
-				await set_trv_values(self, 'valve_position', 0)
-				await asyncio.sleep(60)
-				await set_trv_values(self, 'valve_position', 255)
-				await asyncio.sleep(60)
-				await set_trv_values(self, 'valve_position', 0)
-				await asyncio.sleep(60)
-			else:
-				await set_trv_values(self, 'temperature', 30)
-				await asyncio.sleep(60)
-				await set_trv_values(self, 'temperature', 5)
-				await asyncio.sleep(60)
-				await set_trv_values(self, 'temperature', 30)
-				await asyncio.sleep(60)
-				await set_trv_values(self, 'temperature', 5)
-				await asyncio.sleep(60)
-		
-		self.ignore_states = False
-		
-		# restarting normal heating control immediately
-		await self._async_control_heating()
 	
 	async def _async_control_heating(self):
 		"""main heating control function"""
@@ -922,37 +485,10 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 				else:
 					self._TRV_max_temp = 30
 				
-				# check weather predictions or ambient air temperature if available
-				if self.weather_entity is not None:
-					self.call_for_heat = self.check_weather_prediction()
-				elif self.outdoor_sensor is not None:
-					self.call_for_heat = self.check_ambient_air_temperature()
-				else:
-					self.call_for_heat = True
-				
-				if self.call_for_heat is None:
-					_LOGGER.warning(
-						"better_thermostat: call for heat decision: could not evaluate sensor/weather entity data, force heat on"
-						)
-					self.call_for_heat = True
-				
-				# window open detection and weather detection force turn TRV off
-				if self.window_open and not self.closed_window_triggered:
-					self.last_change = self._hvac_mode
-					self._hvac_mode = HVAC_MODE_OFF
-					self.closed_window_triggered = True
-				elif not self.window_open and self.closed_window_triggered:
-					self._hvac_mode = self.last_change
-					self.closed_window_triggered = False
+				self.call_for_heat = await check_weather(self)
 
-				# check if's summer
-				if self._hvac_mode != HVAC_MODE_OFF and not self.window_open and not self.call_for_heat and not self.load_saved_state:
-					self.last_change = self._hvac_mode
-					self._hvac_mode = HVAC_MODE_OFF
-					self.load_saved_state = True
-				elif self.load_saved_state and self.call_for_heat and not self.window_open:
-					self._hvac_mode = self.last_change
-					self.load_saved_state = False
+				await check_window_state(self)
+				await check_summer_state(self)
 				
 				try:
 					remapped_states = convert_outbound_states(self, self._hvac_mode)
@@ -1014,86 +550,13 @@ class BetterThermostat(ClimateEntity, RestoreEntity, ABC):
 						current_time = datetime.now()
 						if current_time > self.next_valve_maintenance:
 							_LOGGER.debug("better_thermostat: valve_maintenance triggerd")
-							await self.trv_valve_maintenance()
+							await trv_valve_maintenance()
 							self.next_valve_maintenance = datetime.now() + timedelta(days=5)
 				
 				except TypeError as fatal:
 					_LOGGER.debug("better_thermostat entity not ready or device is currently not supported")
 					_LOGGER.debug("fatal %s", fatal)
 					self.ignore_states = False
-	
-	def check_weather_prediction(self):
-		"""
-		Checks configured weather entity for next two days of temperature predictions.
-		@return: True if the maximum forcast temperature is lower than the off temperature; None if not successful
-		"""
-		if self.weather_entity is None:
-			_LOGGER.warning("better_thermostat: weather entity not available.")
-			return None
-		
-		if self.off_temperature is None or not isinstance(self.off_temperature, float):
-			_LOGGER.warning("better_thermostat: off_temperature not set or not a float.")
-			return None
-		
-		try:
-			forcast = self.hass.states.get(self.weather_entity).attributes.get('forecast')
-			if len(forcast) > 0:
-				max_forcast_temp = math.ceil((float(forcast[0]['temperature']) + float(forcast[1]['temperature'])) / 2)
-				_LOGGER.debug("better_thermostat: avg weather temp: %s", max_forcast_temp)
-				return float(max_forcast_temp) < float(self.off_temperature)
-			else:
-				raise TypeError
-		except TypeError:
-			_LOGGER.warning("better_thermostat: no weather entity data found.")
-			return None
-	
-	def check_ambient_air_temperature(self):
-		"""
-		Gets the history for two days and evaluates the necessary for heating.
-		@return: returns True if the average temperature is lower than the off temperature; None if not successful
-		"""
-		if self.outdoor_sensor is None:
-			return None
-		
-		if self.off_temperature is None or not isinstance(self.off_temperature, float):
-			_LOGGER.warning("better_thermostat: off_temperature not set or not a float.")
-			return None
-		
-		try:
-			last_two_days_date_time = datetime.now() - timedelta(days=2)
-			start = dt_util.as_utc(last_two_days_date_time)
-			history_list = history.state_changes_during_period(
-				self.hass, start, dt_util.as_utc(datetime.now()), self.outdoor_sensor
-			)
-			historic_sensor_data = history_list.get(self.outdoor_sensor)
-		except TypeError:
-			_LOGGER.warning("better_thermostat: no outdoor sensor data found.")
-			return None
-		
-		# create a list from valid data in historic_sensor_data
-		valid_historic_sensor_data = []
-		for measurement in historic_sensor_data:
-			if measurement.state is not None:
-				try:
-					valid_historic_sensor_data.append(float(measurement.state))
-				except ValueError:
-					pass
-				except TypeError:
-					pass
-		
-		# remove the upper and lower 5% of the data
-		valid_historic_sensor_data.sort()
-		valid_historic_sensor_data = valid_historic_sensor_data[
-		                             int(len(valid_historic_sensor_data) * 0.05):int(len(valid_historic_sensor_data) * 0.95)]
-		
-		if len(valid_historic_sensor_data) == 0:
-			_LOGGER.warning("better_thermostat: no valid outdoor sensor data found.")
-			return None
-		
-		# calculate the average temperature
-		avg_temp = math.ceil(sum(valid_historic_sensor_data) / len(valid_historic_sensor_data))
-		_LOGGER.debug("better_thermostat: avg outdoor temp: %s", avg_temp)
-		return float(avg_temp) < float(self.off_temperature)
 	
 	@property
 	def _is_device_active(self):
